refactor: tree-shok common.utility
